Standardise drawable state access and split large nested classes out of MainMenu.ButtonSystem
