+++ osu.Game/GameModes/Menu/MenuVisualisation.cs
@@ -1,0 +1,13 @@
+﻿using osu.Framework.Graphics;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace osu.Game.GameModes.Menu
+{
+    internal class MenuVisualisation : Drawable
+    {
+    }
+}
