+++ osu.Game/Overlays/Toolbar.cs
@@ -12,10 +12,11 @@ using System;
 using osu.Framework.Graphics.Transformations;
 using osu.Framework.Timing;
 using osu.Game.GameModes.Play;
+using osu.Framework;
 
 namespace osu.Game.Overlays
 {
-    public class Toolbar : Container
+    public class Toolbar : Container, IStateful<ToolbarState>
     {
         const float height = 50;
 
@@ -25,24 +26,28 @@ namespace osu.Game.Overlays
 
         private ToolbarModeSelector modeSelector;
 
-        public ToolbarState State { get; private set; }
+        private ToolbarState state;
 
-        public void SetState(ToolbarState state, bool instant = false)
+        public ToolbarState State
         {
-            State = state;
-
-            int time = instant ? 0 : 200;
-
-            switch (state)
+            get { return state; }
+            set
             {
-                case ToolbarState.Hidden:
-                    MoveToY(-Size.Y, time, EasingTypes.InQuint);
-                    FadeOut(time);
-                    break;
-                case ToolbarState.Visible:
-                    MoveToY(0, time, EasingTypes.OutQuint);
-                    FadeIn(time);
-                    break;
+                state = value;
+
+                const int transition_time = 200;
+
+                switch (state)
+                {
+                    case ToolbarState.Hidden:
+                        MoveToY(-Size.Y, transition_time, EasingTypes.InQuint);
+                        FadeOut(transition_time);
+                        break;
+                    case ToolbarState.Visible:
+                        MoveToY(0, transition_time, EasingTypes.OutQuint);
+                        FadeIn(transition_time);
+                        break;
+                }
             }
         }
